add test that this project's license detected as MIT
